requirements.txt
tensorflow
torch
transformers
faiss-cpu
flask
ollama
fastapi
uvicorn
pillow
sentence-transformers
reportlab